chore: move generateHtml
